Add multiple methods in Endpoint
