+++ modules/lepus-router/src/main/scala/lepus/router/internal/ExtensionMethods.scala
@@ -8,36 +8,6 @@ import lepus.router.http.*
 
 trait ExtensionMethods extends lepus.core.internal.ExtensionMethods:
 
-  extension (endpoint: Endpoint[?])
-    def recursiveEndpoints[T](pf: PartialFunction[Endpoint[?], Vector[T]]): Vector[T] =
-      endpoint match
-        case Endpoint.Pair(left, right)          => left.recursiveEndpoints(pf) ++ right.recursiveEndpoints(pf)
-        case r: Endpoint[?] if pf.isDefinedAt(r) => pf(r)
-        case _                                   => Vector.empty
-
-    def asVector(): Vector[Endpoint[?]] =
-      recursiveEndpoints {
-        case e: Endpoint[?] => Vector(e)
-      }
-
-    def toPath: String = "/" + asVector()
-      .map {
-        case e: Endpoint.FixedPath[?] => e.name
-        case e: Endpoint.Path[?]      => s"{${ e.name }}"
-        case _                        => ""
-      }
-      .mkString("/")
-
-    def isPath: Boolean =
-      endpoint match
-        case _: Endpoint.Path[?] => true
-        case _                   => false
-
-    def isQueryParam: Boolean =
-      endpoint match
-        case _: Endpoint.Query[?] => true
-        case _                    => false
-
   // see https://github.com/scala/bug/issues/12186
   extension [T](v: Vector[T])
     def headAndTail: Option[(T, Vector[T])] = if v.isEmpty then None else Some((v.head, v.tail))
